remote promo data support
