+++ dev.env
@@ -5,6 +5,7 @@ LISTEN_IP=0.0.0.0
 LISTEN_PORT=8080
 REQUIRE_SESSION_TOKEN=
 ALLOW_NON_MODINFO=
+PROMOS_REMOTE=
 
 # not used for dev
 SSL_KEY_PATH=
